???
so turns out that i somehow don't need the position variables to be pointers, because it works this way as well··· but this is actually weird 'cos i do remember it not working if i did that, and that was when i had to specify `mvspr()` of my positions
maybe the behavior is different on Clang?
also tried to have `shoot()` not need to return anything as it looked quite hacky, but then for that one i wasn't able to make it work even by setting the HP parameter as `uintptr_t`: regardless of the type i used, the HP would only go down by one once something got shot, but then it'd stay frozen at that value instead of remaining to go down as more bullets shot it, so neither you nor your opponent would win anything because both HPs couldn't get down any further for whatever fucked reason C has
also remove unecessary wasteful iteration in `gameover()`

+++ sprites.h
@@ -4,7 +4,7 @@
 
 // a sprite here has a
 typedef struct SPRITE {
-	uintptr_t y, x; // starting and current position
+	unsigned short y, x; // starting and current position
 	const unsigned short h, w; // height & width
 	unsigned char hit[2]; // hitbox (0: y, 1: x)
 	unsigned char hp; // health
